ai-services/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-core
pydantic
python-dotenv